changes requested + linter
