+++ frontend/src/components/Projects/ProjectList.js
@@ -170,12 +170,12 @@ const ProjectList = () => {
   };
 
   const getLogoURL = (item) => {
-      if(item && item.name) {
-          let cleanSeed = _.snakeCase(item.name);
-          return 'https://picsum.photos/seed/' + cleanSeed + '/40';
-      }else{
-          return 'https://picsum.photos/40';
-      }
+    if(item && item.name) {
+      let cleanSeed = _.snakeCase(item.name);
+      return 'https://picsum.photos/seed/' + cleanSeed + '/40';
+    }else{
+      return 'https://picsum.photos/40';
+    }
   };
 
   const DataList = ({data}) => <List
