+++ frontend/src/components/TimeEntry/AddEntryForm.less
@@ -16,7 +16,7 @@
 
 @import '../../main.less';
 
-.tk-radio-disabled {
+.tk-radio {
   .ant-radio-disabled {
     .ant-radio-inner {
       background-color: @tk_DarkGray !important;
